# app/openapiDefinitions/open_api_spec_0.1.0.yaml
openapi: 3.0.0
info:
  title: UAA
  version: 0.1.0
paths:
...
  /users:
    post:
      operationId: createUser
      summary: Create a new user
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        201:
          description: The user has been created
        400:
          $ref: "#/components/responses/BadRequest"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    CreateUserRequest:
      type: object
      required:
        - username
        - email
        - password
        - roles
      properties:
        username:
          type: string
          description: Username of the user
        email:
          type: string
          description: Email of the user
        password:
          type: string
          description: Password of the user
        roles:
          type: array
          description: Roles to assign to the user
          items:
            type: string
            format: uuid
            description: Identifier of the role
...
    BadRequestSchema:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Message with detailed description about the problem
...
  responses:
    BadRequest:
      description: There is a failure in the request format, expected headers, or the payload can't be unmarshalled
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BadRequestSchema"
...